pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xlreason-iqgp"
version = "0.1.0"
description = "Bilingual reasoning via an interlingua QID-Graph planner"
authors = [{name = "ErrDivine"}]
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
